re-write without angstrom and faraday. (#41)
This is a full re-write to write the core library without relying on angstrom and faraday. The core library is written using `Core.Iobuf.t` and the parsing layer has been made simpler to write/maintain and performs more checks that each parser consumes exactly `n` bytes where `n` is the frame length specified in Postgres backend message's frame header. Removing angstrom/faraday also allowed writing a simpler interface for writing IO wrappers that can drive the parser and serializer. Some additions since the previous commit is that the library now has all the building blocks to support `SCRAM-SHA-256` auth mechanism which is the default starting Postgresql 14. MD5 and Cleartext auth mechanisms are still supported as well.

I'm also taking this as an opportunity to explore `eio` and `postgres_eio` is the only IO driver that's being worked on at the moment.

closes: https://github.com/anuragsoni/postgres-protocol/issues/11